backend/requirements.txt
Flask==3.1.2
requests==2.32.5
PyJWT==2.11.0
functions-framework==3.10.0
